Processamento/ContaCapitalProcessamentoEJB/src/test/java/br/com/sicoob/sisbr/cca/processamento/negocio/servicos/interceptors/ProcessamentoInternacionalizacaoInterceptorTest.java
package br.com.sicoob.sisbr.cca.processamento.negocio.servicos.interceptors;

import org.junit.Assert;
import org.junit.BeforeClass;
import org.junit.Test;

import br.com.bancoob.negocio.contexto.InformacoesUsuario;

public class ProcessamentoInternacionalizacaoInterceptorTest {

	@BeforeClass
	public static void setUpClass() {
		InformacoesUsuario.INSTANCIA.set(new InformacoesUsuario());
	}
	
	@Test
	public void deveInstanciarInterceptor() {
		ProcessamentoInternacionalizacaoInterceptor interceptor = 
				new ProcessamentoInternacionalizacaoInterceptor();
		Assert.assertNotNull(interceptor);
		Assert.assertNotNull(interceptor.getResourceBundle());
	}
	
}
